+++ scripts/combat.js
@@ -1,4 +1,5 @@
 import { dialog } from "./dialog/dialog.js";
+import { actors } from "./actors/actors.js";
 import { playlists } from "./playlists.js";
 import { numbers } from "./utility/numbers.js";
 import { ranged } from "./weapons/ranged.js";
@@ -9,6 +10,9 @@ import { bloodyAxe } from "./items/weapons/bloodyAxe.js";
 import { mounts } from "./mounts/mounts.js";
 import { tagging } from "./utility/tagging.js";
 import { versioning } from "./versioning.js";
+import { folders } from "./folders/folders.js";
+import { importFromCompedium } from "./gm/gmFunctions.js";
+import { items } from "./items/items.js";
 
 export let combat = {
     "applyAdhocDamage": foundry.utils.debounce(applyAdhocDamage, 250),
@@ -34,14 +38,14 @@ export let combat = {
 
 const WALLS_BLOCK_RANGE = "optionalRules.wallsBlockRange";
 async function toggleWallsBlockRanged() {
-    
+
     let midiConfig = game.settings.get("midi-qol", "ConfigSettings");
     const currentSetting = foundry.utils.getProperty(midiConfig, WALLS_BLOCK_RANGE);
     let wallsBlockRange = currentSetting != "none";
     let messageContent = wallsBlockRange ? "Walls will not block range." : "Walls block range.";
     if (wallsBlockRange) {
         await game.user.setFlag(sdndConstants.MODULE_ID, WALLS_BLOCK_RANGE, currentSetting);
-    } 
+    }
     const newSetting = wallsBlockRange ?
         "none" :
         game.user.getFlag(sdndConstants.MODULE_ID, WALLS_BLOCK_RANGE) ?? "center";
@@ -51,6 +55,18 @@ async function toggleWallsBlockRanged() {
         content: messageContent,
         whisper: ChatMessage.getWhisperRecipients('GM'),
     });
+}
+
+function getSortedNames(targets) {
+    let sortedNames = targets.map(t => t.name).sort();
+    let targetNames = sortedNames.slice(0, sortedNames.length - 1).join(`, `);
+    if (sortedNames.length > 1) {
+        targetNames += ` and ${sortedNames[sortedNames.length - 1]} have`;
+    }
+    else {
+        targetNames = `${sortedNames[0]} has`;
+    }
+    return targetNames;
 }
 
 function getWeaponDamageTypes(weapon) {
@@ -120,6 +136,60 @@ async function onHealed(actor, changes, update, userId) {
     await mounts.hooks.onHealed(actor, changes, update, userId);
 }
 
+async function createAdhocItem(adhocDamageType, damageType, damageDice, diceCount, saveData, targets) {
+    let adhocActor = await actors.ensureActor("Adhoc Damage", sdndConstants.PACKS.COMPENDIUMS.ACTOR.TEMP, folders.Actor.TempActors.name);
+    if (!adhocActor) {
+        adhocActor = await importFromCompedium("Actor", sdndConstants.PACKS.COMPENDIUMS.ACTOR.TEMP, "7vOYBKZ0yWut0zCw", tempActorsFolder.name);
+    }
+    let item = adhocActor.items.find(i => i.name === adhocDamageType);
+    if (!item) {
+        await ui.notifications.error("Could not create adhoc damage synthetic item!");
+        return null;
+    }
+    const dnd5eDamageType = CONFIG.DND5E.damageTypes[damageType];
+    const damageImage = `modules/stroud-dnd-helpers/images/icons/damageTypes/${damageType}_damage.webp`;
+    let activity = item.system.activities.contents[0];
+    let parentActor = item.parent;
+    await parentActor.updateEmbeddedDocuments(Item.name, [
+        {
+            "_id": item._id,
+            "system": {
+                description: {
+                    "chat": `<img src='${damageImage}' style='border: 0!important'/>`
+                }
+            }
+        }
+    ]);
+    const flavor = `${getSortedNames(targets)} been struck with <span style='${dnd5eDamageType.color}'>${dnd5eDamageType.label}</span> damage!`;
+    let changes = {
+        name: dnd5eDamageType.label,
+        description: {
+            chatFlavor: flavor
+        },
+        damage: {
+            parts: [
+                {
+                    denomination: numbers.toNumber(damageDice),
+                    number: numbers.toNumber(diceCount),
+                    types: [damageType]
+                }
+            ]
+        }
+    };
+    if (saveData) {
+        changes.damage.onSave = saveData.damageOnSave;
+        changes.save = {
+            ability: [saveData.ability],
+            dc: {
+                formula: `${saveData.dc}`,
+                value: saveData.dc
+            }
+        }
+    }
+    await activity.update(changes);
+    return item;
+}
+
 // apply adhoc damage to selected tokens
 async function applyAdhocDamage() {
     if (!game.user.isGM) {
@@ -127,12 +197,14 @@ async function applyAdhocDamage() {
         return;
     }
     let adHocDamage = {
+        getAdhocDamageType: async function _getAdhocDamageType() {
+            return await dialog.createButtonDialog("Adhoc Damage Type", sdndConstants.BUTTON_LISTS.ADHOC_DAMAGE_TYPE, 'column');
+        },
         getDamageType: async function _getDamageType() {
             return await dialog.createButtonDialog("Select Damage Type", sdndConstants.BUTTON_LISTS.DAMAGE_5E, 'column');
         },
         getDamageDice: async function _getDamageDice() {
-            let diceButtons = ['d4', 'd6', 'd8', 'd10', 'd20', 'd100'];
-            return await dialog.createButtonDialog("Select Damage Dice", diceButtons.map(v => ({ label: v, value: v })), 'column');
+            return await dialog.createButtonDialog("Select Damage Dice", sdndConstants.BUTTON_LISTS.DICE_TYPES, 'column');
         },
         getDiceCount: async function _getDiceCount() {
             let numberButtons = [];
@@ -141,16 +213,30 @@ async function applyAdhocDamage() {
             }
             return await dialog.createButtonDialog("How Many Dice?", numberButtons, 'row');
         },
-        getSortedNames: function _getSortedNames(targets) {
-            let sortedNames = targets.map(t => t.name).sort();
-            let targetNames = sortedNames.slice(0, sortedNames.length - 1).join(`, `);
-            if (sortedNames.length > 1) {
-                targetNames += ` and ${sortedNames[sortedNames.length - 1]} have`;
-            }
-            else {
-                targetNames = `${sortedNames[0]} has`;
-            }
-            return targetNames;
+        "getSortedNames": getSortedNames,
+        getSaveAbility: async function _getSaveAbility() {
+            const abilities = Object.values(CONFIG.DND5E.abilities).map(a => ({ label: a.label, value: a.abbreviation }));
+            return await dialog.createButtonDialog("Select Save Ability", abilities, "column");
+        },
+        getSaveDC: async function _getSaveDC() {
+            const values = [5, 10, 15, 20, 25, 30].map(v => ({
+                label: v,
+                value: v
+            }));
+            return await dialog.createButtonDialog("Select Save DC", values, "column");
+        },
+        getDamageOnSave: async function _getDamageOnSave() {
+            const values = [
+                {
+                    label: "Half Damage",
+                    value: "half"
+                },
+                {
+                    label: "No Damage",
+                    value: "none"
+                }
+            ];
+            return await dialog.createButtonDialog("Select Damage on Save", values, "column");
         }
     };
 
@@ -162,28 +248,55 @@ async function applyAdhocDamage() {
         ui.notifications.notify(`No valid tokens selected!`);
         return;
     }
+    let saveData = null;
+    let adhocDamageType = await adHocDamage.getAdhocDamageType();
+    if (!adhocDamageType) {
+        return;
+    }
+
+    if (adhocDamageType === "Damage") {
+        const saveAbility = await adHocDamage.getSaveAbility();
+        if (!saveAbility) {
+            return;
+        }
+        const saveDC = await adHocDamage.getSaveDC();
+        if (!saveDC) {
+            return;
+        }
+        const damageOnSave = await adHocDamage.getDamageOnSave();
+        if (!damageOnSave) {
+            return;
+        }
+        saveData = {
+            ability: saveAbility,
+            dc: saveDC,
+            damageOnSave: damageOnSave
+        };
+    }
 
     let damageType = await adHocDamage.getDamageType();
     if (!damageType) {
         return;
     }
     let damageDice = await adHocDamage.getDamageDice();
-    if (!damageDice) {s
+    if (!damageDice) {
         return;
     }
     let diceCount = await adHocDamage.getDiceCount();
     if (!diceCount) {
         return;
     }
-    const damageRoll = await new Roll(`${diceCount}${damageDice}[${damageType}]`).evaluate();
     const dnd5eDamageType = CONFIG.DND5E.damageTypes[damageType];
     let color = dnd5eDamageType.color;
     if (color) {
         color = `color:${color}`;
     }
-    damageRoll.toMessage({ flavor: `${adHocDamage.getSortedNames(targets)} been struck with <span style='${color}'>${dnd5eDamageType.label}</span> damage!` });
-    let autoApplyAdhocDamage = sdndSettings.AutoApplyAdhocDamage.getValue();
-    await MidiQOL.applyTokenDamage([{ type: `${damageType}`, damage: damageRoll.total }], damageRoll.total, new Set(targets), null, new Set(), { forceApply: autoApplyAdhocDamage });
+    await applyDamage(adhocDamageType, damageType, damageDice, diceCount, saveData, targets);
+}
+
+async function applyDamage(adhocDamageType, damageType, damageDice, diceCount, saveData, targets) {
+    const item = await createAdhocItem(adhocDamageType, damageType, damageDice, diceCount, saveData, targets);
+    return await chrisPremades.utils.workflowUtils.syntheticItemRoll(item, targets, { userId: game.userId });
 }
 
 const damageMatch = /\[(?<damage>\w+)\]/;
@@ -200,12 +313,12 @@ async function applyAttackers(attackerCount, rollFormula, hitModifier, autoApply
         ui.notifications.notify(`No valid tokens selected!`);
         return;
     }
-    let results = Object.fromEntries(canvas.tokens.controlled.map(t =>  [t.id, { 'name': t.name, 'hits': 0, 'damage': 0 }]));
+    let results = Object.fromEntries(canvas.tokens.controlled.map(t => [t.id, { 'name': t.name, 'hits': 0, 'damage': 0 }]));
     let damageType = damageMatch.exec(rollFormula)?.groups["damage"] ?? 'slashing';
     for (let i = 0; i < attackerCount; i++) {
         let currentTarget = targets[i % targets.length];
         const hitRoll = await new Roll(`1d20+${hitModifier}`).evaluate();
-        if (hitRoll.total <= (currentTarget?.actor?.system?.attributes?.ac?.value ?? 10)){
+        if (hitRoll.total <= (currentTarget?.actor?.system?.attributes?.ac?.value ?? 10)) {
             continue;
         }
         const damageRoll = await new Roll(rollFormula).evaluate();