+++ scripts/constants.js
@@ -64,6 +64,46 @@ export const sdndConstants = {
                 "label": "🌩 Thunder",
                 "value": "thunder"
             }
+        ],
+        "DICE_TYPES": [
+            {
+                "label": "d4",
+                "value": "4"
+            },
+            {
+                "label": "d6",
+                "value": "6"
+            },
+            {
+                "label": "d8",
+                "value": "8"
+            },
+            {
+                "label": "d10",
+                "value": "10"
+            },
+            {
+                "label": "d12",
+                "value": "12"
+            },
+            {
+                "label": "d20",
+                "value": "20"
+            },
+            {
+                "label": "d100",
+                "value": "100"
+            }
+        ],
+        "ADHOC_DAMAGE_TYPE": [
+            {
+                "label": "Allow Saves",
+                "value": "Damage"
+            },
+            {
+                "label": "No Saves",
+                "value": "Unavoidable Damage"
+            }
         ]
     },
     "FEATURES": {