+++ scripts/folders/folders.js
@@ -1,24 +1,29 @@
 import { gmFunctions } from "../gm/gmFunctions.js";
 
+const tempActors = {
+    "name": "SND Temporary Actors",
+    "type": "Actor",
+    "folders": [],
+    "ensureFolder": async function _ensureFolder(parentID) {
+        return await folders.ensureFolder(this.name, this.type, parentID);
+    }
+};
+
 export let folders = {
     "Actor": {
-        "TempActors": {
-            "name": "SND Temporary Actors",
-            "type": "Actor",
-            "folders": [],
-            "ensureFolder": async () => ensureFolder(this.name, this.type, null)
-        }
+        "TempActors": tempActors
     },
     "ensureFolder": ensureFolder,
     "childOfFolder": childOfFolder
 };
+
 
 async function createFolder(folderName, folderType, parentID) {
     return await Folder.create({ "name": folderName, "type": folderType, "parent": parentID });
 }
 
 async function ensureFolder(folderName, folderType, parentID) {
-    let folder = await game.folders.find(f => f.name == folderName && f.folder?.id == parentID);
+    let folder = game.folders.find(f => f.name == folderName && f.folder?.id == parentID);
 
     if (!folder) {
         folder = await createFolder(folderName, folderType, parentID);